+++ src/runtime/server/lib/oauth/facebook.ts
@@ -8,6 +8,7 @@ import {
 } from 'h3'
 import { withQuery } from 'ufo'
 import { defu } from 'defu'
+import { handleAccessTokenErrorResponse, handleMissingConfiguration } from '../utils'
 import { useRuntimeConfig } from '#imports'
 import type { OAuthConfig } from '#auth-utils'
 
@@ -86,13 +87,7 @@ export function oauthFacebookEventHandler({
     }
 
     if (!config.clientId) {
-      const error = createError({
-        statusCode: 500,
-        message:
-          'Missing NUXT_OAUTH_FACEBOOK_CLIENT_ID or NUXT_OAUTH_FACEBOOK_CLIENT_SECRET env variables.',
-      })
-      if (!onError) throw error
-      return onError(event, error)
+      return handleMissingConfiguration(event, 'facebook', ['clientId'], onError)
     }
 
     const redirectURL = config.redirectURL || getRequestURL(event).href
@@ -120,14 +115,9 @@ export function oauthFacebookEventHandler({
         code: query.code,
       },
     })
+
     if (tokens.error) {
-      const error = createError({
-        statusCode: 401,
-        message: `Facebook login failed: ${tokens.error || 'Unknown error'}`,
-        data: tokens,
-      })
-      if (!onError) throw error
-      return onError(event, error)
+      return handleAccessTokenErrorResponse(event, 'facebook', tokens, onError)
     }
 
     const accessToken = tokens.access_token