+++ src/runtime/server/lib/oauth/google.ts
@@ -1,13 +1,13 @@
 import type { H3Event } from 'h3'
 import {
   eventHandler,
-  createError,
   getQuery,
   getRequestURL,
   sendRedirect,
 } from 'h3'
 import { withQuery, parsePath } from 'ufo'
 import { defu } from 'defu'
+import { handleAccessTokenErrorResponse, handleMissingConfiguration } from '../utils'
 import { useRuntimeConfig } from '#imports'
 import type { OAuthConfig } from '#auth-utils'
 
@@ -79,12 +79,7 @@ export function oauthGoogleEventHandler({
     const { code } = getQuery(event)
 
     if (!config.clientId) {
-      const error = createError({
-        statusCode: 500,
-        message: 'Missing NUXT_OAUTH_GOOGLE_CLIENT_ID env variables.',
-      })
-      if (!onError) throw error
-      return onError(event, error)
+      return handleMissingConfiguration(event, 'google', ['clientId'], onError)
     }
 
     const redirectURL = config.redirectURL || getRequestURL(event).href
@@ -121,15 +116,7 @@ export function oauthGoogleEventHandler({
       return { error }
     })
     if (tokens.error) {
-      const error = createError({
-        statusCode: 401,
-        message: `Google login failed: ${
-          tokens.error?.data?.error_description || 'Unknown error'
-        }`,
-        data: tokens,
-      })
-      if (!onError) throw error
-      return onError(event, error)
+      return handleAccessTokenErrorResponse(event, 'google', tokens, onError)
     }
 
     const accessToken = tokens.access_token