+++ src/runtime/server/lib/oauth/linkedin.ts
@@ -1,7 +1,8 @@
 import type { H3Event, H3Error } from 'h3'
-import { eventHandler, createError, getQuery, getRequestURL, sendRedirect } from 'h3'
+import { eventHandler, getQuery, getRequestURL, sendRedirect } from 'h3'
 import { withQuery, parseURL, stringifyParsedURL } from 'ufo'
 import { defu } from 'defu'
+import { handleAccessTokenErrorResponse, handleMissingConfiguration } from '../utils'
 import { useRuntimeConfig } from '#imports'
 
 export interface OAuthLinkedInConfig {
@@ -66,12 +67,7 @@ export function oauthLinkedInEventHandler({ config, onSuccess, onError }: OAuthC
     const { code } = getQuery(event)
 
     if (!config.clientId || !config.clientSecret) {
-      const error = createError({
-        statusCode: 500,
-        message: 'Missing NUXT_OAUTH_LINKEDIN_CLIENT_ID or NUXT_OAUTH_LINKEDIN_CLIENT_SECRET env variables.',
-      })
-      if (!onError) throw error
-      return onError(event, error)
+      return handleMissingConfiguration(event, 'linkedin', ['clientId', 'clientSecret'], onError)
     }
 
     const redirectURL = config.redirectURL || getRequestURL(event).href
@@ -119,13 +115,7 @@ export function oauthLinkedInEventHandler({ config, onSuccess, onError }: OAuthC
       return { error }
     })
     if (tokens.error) {
-      const error = createError({
-        statusCode: 401,
-        message: `LinkedIn login failed: ${tokens.error?.data?.error_description || 'Unknown error'}`,
-        data: tokens,
-      })
-      if (!onError) throw error
-      return onError(event, error)
+      return handleAccessTokenErrorResponse(event, 'linkedin', tokens, onError)
     }
 
     const accessToken = tokens.access_token