+++ src/runtime/server/lib/utils.ts
@@ -1,0 +1,37 @@
+import type { H3Event } from 'h3'
+
+import { snakeCase, upperFirst } from 'scule'
+import type { OnError, OAuthProvider } from '#auth-utils'
+import { createError } from '#imports'
+
+/**
+ * Handle OAuth access token error response
+ *
+ * @see https://datatracker.ietf.org/doc/html/rfc6749#section-5.2
+ */
+// TODO: waiting for https://github.com/atinux/nuxt-auth-utils/pull/140
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+export function handleAccessTokenErrorResponse(event: H3Event, oauthProvider: OAuthProvider, oauthError: any, onError?: OnError) {
+  const message = `${upperFirst(oauthProvider)} login failed: ${oauthError.error_description || oauthError.error || 'Unknown error'}`
+
+  const error = createError({
+    statusCode: 401,
+    message,
+    data: oauthError,
+  })
+
+  if (!onError) throw error
+  return onError(event, error)
+}
+
+export function handleMissingConfiguration(event: H3Event, provider: OAuthProvider, missingKeys: string[], onError?: OnError) {
+  const environmentVariables = missingKeys.map(key => `NUXT_OAUTH_${provider.toUpperCase()}_${snakeCase(key).toUpperCase()}`)
+
+  const error = createError({
+    statusCode: 500,
+    message: `Missing ${environmentVariables.join(' or ')} env ${missingKeys.length > 1 ? 'variables' : 'variable'}.`,
+  })
+
+  if (!onError) throw error
+  return onError(event, error)
+}