+++ src/runtime/types/index.ts
@@ -1,3 +1,3 @@
 export type { User, UserSession, UserSessionRequired, UserSessionComposable } from './session'
-export type { OAuthConfig } from './oauth-config'
+export type { OAuthConfig, OAuthProvider, OnError } from './oauth-config'
 export type { WebauthnComposable } from './webauthn'